Move DAM blocks to dam/ folder
