+++ packages/api/cms-api/src/blocks/createTextImageBlock.ts
@@ -15,8 +15,8 @@ import {
 import { Type } from "class-transformer";
 import { IsEnum, IsString } from "class-validator";
 
+import { PixelImageBlock } from "../dam/blocks/pixel-image.block";
 import { IsAllowedImageAspectRatio } from "../dam/images/validators/is-allowed-aspect-ratio.validator";
-import { PixelImageBlock } from "./PixelImageBlock";
 
 export enum ImagePosition {
     Left = "left",
