+++ packages/api/cms-api/src/dam/blocks/dam-video-block-transformer.service.ts
@@ -1,8 +1,8 @@
 import { BlockContext, BlockTransformerServiceInterface } from "@comet/blocks-api";
 import { Injectable } from "@nestjs/common";
 
-import { FilesService } from "../dam/files/files.service";
-import { DamScopeInterface } from "../dam/types";
+import { FilesService } from "../files/files.service";
+import { DamScopeInterface } from "../types";
 import { DamVideoBlockData } from "./dam-video.block";
 
 type TransformReturn = {
