+++ packages/api/cms-api/src/dam/blocks/pixel-image.block.ts
@@ -12,12 +12,12 @@ import {
 import { Type } from "class-transformer";
 import { IsNotEmpty, IsOptional, IsString, ValidateNested } from "class-validator";
 
-import { FocalPoint } from "../dam/common/enums/focal-point.enum";
-import { FILE_ENTITY } from "../dam/files/entities/file.entity";
-import { FilesService } from "../dam/files/files.service";
-import { ImageCropAreaInput } from "../dam/images/dto/image-crop-area.input";
-import { ImageCropArea } from "../dam/images/entities/image-crop-area.entity";
-import { ImagesService } from "../dam/images/images.service";
+import { FocalPoint } from "../common/enums/focal-point.enum";
+import { FILE_ENTITY } from "../files/entities/file.entity";
+import { FilesService } from "../files/files.service";
+import { ImageCropAreaInput } from "../images/dto/image-crop-area.input";
+import { ImageCropArea } from "../images/entities/image-crop-area.entity";
+import { ImagesService } from "../images/images.service";
 import { PixelImageBlockTransformerService } from "./pixel-image-block-transformer.service";
 
 // @TODO: make factory to support flexible validation
