+++ packages/api/cms-api/src/dam/dam.module.ts
@@ -3,9 +3,9 @@ import { DynamicModule, Global, Module, Type, ValueProvider } from "@nestjs/comm
 import { TypeMetadataStorage } from "@nestjs/graphql";
 
 import { BlobStorageModule, defaultDamAcceptedMimetypes, DependentsResolverFactory } from "..";
-import { DamVideoBlockTransformerService } from "../blocks/dam-video-block-transformer.service";
-import { PixelImageBlockTransformerService } from "../blocks/pixel-image-block-transformer.service";
-import { SvgImageBlockTransformerService } from "../blocks/svg-image-block-transformer.service";
+import { DamVideoBlockTransformerService } from "./blocks/dam-video-block-transformer.service";
+import { PixelImageBlockTransformerService } from "./blocks/pixel-image-block-transformer.service";
+import { SvgImageBlockTransformerService } from "./blocks/svg-image-block-transformer.service";
 import { ScaledImagesCacheService } from "./cache/scaled-images-cache.service";
 import { DamConfig } from "./dam.config";
 import { DAM_CONFIG, DAM_FILE_VALIDATION_SERVICE, IMGPROXY_CONFIG } from "./dam.constants";
